requirements.txt
# -------------------------------------
# Plan It
# Web Service
# 'requirements.txt'
# Contact: planit.app.dev@gmail.com
# -------------------------------------

Flask
Flask-Sessionstore
pymongo
python-dotenv
marshmallow
